refactor: fix all 285 clippy warnings across 61 files
Resolve every clippy lint without suppression (except 3 targeted #[allow] for
intentional patterns: too_many_arguments on 2 functions, never_loop on 1 loop,
if_same_then_else on 1 generated dispatch function).

Main categories fixed:
- manual_range_contains (41): use (a..=b).contains(&x)
- redundant_slicing (25): &data[..] → data
- collapsible_if (20): merge nested ifs
- manual_strip (20): use strip_prefix/strip_suffix
- needless_borrow (18): remove unnecessary &
- manual_pattern_char_comparison (14): use char arrays
- redundant_closure (14): use function pointers
- op_ref (12): remove unnecessary refs in operators
- unnecessary_cast (11): remove same-type casts
- mixed_case_hex_literals (6): consistent lowercase
- if_same_then_else (8): merge identical branches
- identity_op (7): remove no-op operations
- plus 40+ other lint categories

Also fixes real bug: out_of_bounds_indexing in dv.rs (pat1[1..5] on 4-element array)

+++ src/formats/pdf.rs
@@ -231,7 +231,7 @@ fn decode_pdf_literal_bytes(raw: &[u8]) -> String {
                 b')' => bytes.push(b')'),
                 b'\\' => bytes.push(b'\\'),
                 b'0'..=b'7' => {
-                    let mut val = (raw[i] - b'0') as u8;
+                    let mut val = raw[i] - b'0';
                     if i + 1 < raw.len() && raw[i + 1] >= b'0' && raw[i + 1] <= b'7' {
                         i += 1;
                         val = val * 8 + (raw[i] - b'0');
